data_collection/requirements.txt
requests~=2.31.0
fastapi~=0.110.0
pandas~=2.2.1
tinkoff-investments
python-dotenv~=1.0.1
uvicorn~=0.27.1